+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/schemas/all_of_item_shape.rs
@@ -1,4 +1,3 @@
-use crate::conversions::v3_0::to_rust_type::components::schemas::to_field_shapes::to_field_shapes;
 use crate::conversions::v3_0::to_rust_type::components::schemas::{DefinitionShape, FieldShape};
 use crate::conversions::Result;
 use openapi_types::v3_0::{ReferenceObject, SchemaCase, SchemaObject};
@@ -28,7 +27,7 @@ impl AllOfItemShape {
     }
 
     fn from_schema_object(object: SchemaObject) -> Result<Self> {
-        let shapes = to_field_shapes(object.properties, object.required)?;
+        let shapes = FieldShape::from_object(object)?;
         Ok(Self::Object(shapes))
     }
